server/pyproject.toml
[tool.poetry]
name = "boilerplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
celery = "^5.4.0"
redis = "^5.0.7"
flower = "^2.0.1"
django-cors-headers = "^4.4.0"
pytest = "^8.2.2"
tornado = "^6.4.1"
pillow = "^10.3.0"
django-rest-auth = "^0.9.5"
ipython = "^8.25.0"
django-allauth = "^0.63.3"
django-extensions = "^3.2.3"
django-extra-fields = "^3.0.2"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
dj-rest-auth = "^6.0.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
